+++ src/component/main/Banner.tsx
@@ -1,4 +1,4 @@
-import { EmptyDiv40h } from '@/styles/CommomStyle/Common.styled';
+import { Gap } from '@/styles/CommomStyle/Common.styled';
 import theme from '@/styles/theme';
 import styled from '@emotion/styled';
 
@@ -38,7 +38,7 @@ const Banner = () => {
         <BannerSub>카카오테크 캠퍼스 3기여러분</BannerSub>
         <BannerMain>프론트엔드 2단계 과제 화이팅! 🎉</BannerMain>
       </BannerTextWrapper>
-      <EmptyDiv40h/>
+      <Gap height={40} />
     </BannerSection>
   );
 };
