+++ src/component/main/GiftRankingList.tsx
@@ -7,11 +7,12 @@ import { useAuth } from '@/context/AuthContext';
 import { useNavigate } from 'react-router-dom';
 import type { ProductItem } from '@/type/GiftAPI/product';
 import type { RankType, TargetType } from '@/type/giftRanking';
-import { CentorAlignDiv240, EmptyDiv16h } from '@/styles/CommomStyle/Common.styled';
+import { CentorAlignDiv240, Gap } from '@/styles/CommomStyle/Common.styled';
 import { BrandImage, Price, ProductCard, ProductGrid, ProductImage, ProductInfo } from '@/styles/CommomStyle/ProductList';
 import useFetchFromUrlT from '@/hook/useFetchFromUrlT';
 import Loading from '../Loading';
 import { baseRankingUrl } from '@/constant/api';
+import { getFromUrl } from '@/utils/getFromUrl';
 
 
 
@@ -27,7 +28,7 @@ const GIFTLENGTH = 6;
 const GiftRankingList = ({ targetType, rankType }: GiftRankingListProps) => {
     const RankingUrl = `${baseRankingUrl}?targetType=${targetType}&rankType=${rankType}`
     const [isExpanded, setIsExpanded] = useState(false);
-    const { item, loading, error } = useFetchFromUrlT<[]>(RankingUrl, []);
+    const { item, loading, error } = useFetchFromUrlT<[]>(RankingUrl, getFromUrl, []);
     const { user } = useAuth();
     const navigate = useNavigate();
 
@@ -71,7 +72,7 @@ const GiftRankingList = ({ targetType, rankType }: GiftRankingListProps) => {
                     </ProductCard>
                 ))}
             </ProductGrid>
-            <EmptyDiv16h />
+            <Gap height={16}  />
             <LoadMoreButtonDiv>
                 {item.length > GIFTLENGTH && (
                     <LoadMoreButton onClick={() => setIsExpanded((prev) => !prev)}>
@@ -82,7 +83,7 @@ const GiftRankingList = ({ targetType, rankType }: GiftRankingListProps) => {
                 )}
             </LoadMoreButtonDiv>
 
-            <EmptyDiv16h />
+            <Gap height={16}  />
         </>
     )
 };
