+++ src/hook/useInfiniteFetchFromUrlT.ts
@@ -1,8 +1,7 @@
 import type { ApiResponse } from "@/type/GiftAPI/product";
-import { getFromUrl } from "@/utils/getFromUrl";
 import { useEffect, useState } from "react";
 
-function useFetchFromUrlT<T>(url: string, defaultT: T) {
+function useInfiniteFetchFromUrlT<T>( url: string, fetchFn : (url : string) => Promise<ApiResponse<T>>, defaultT: T) {
     const [item, setItem] = useState<T>(defaultT);
     const [loading, setLoading] = useState<boolean>(true);
     const [error, setError] = useState<Error | null>(null);
@@ -10,10 +9,9 @@ function useFetchFromUrlT<T>(url: string, defaultT: T) {
 
     useEffect(() => {
         let isMounted = true;
-        setLoading(true);
         const fetchData = async () => {
             try {
-                const newItem = await getFromUrl<ApiResponse<T>>(url);
+                const newItem = await fetchFn(url);
 
 
                 if (!isMounted) return;
@@ -22,7 +20,6 @@ function useFetchFromUrlT<T>(url: string, defaultT: T) {
 
             } catch (error) {
                 setError(error as Error);
-                throw new Error(`${url} 데이터 Fetch 실패,  ${(error as Error).message}`);
             } finally {
                 setLoading(false);
             }
@@ -38,4 +35,4 @@ function useFetchFromUrlT<T>(url: string, defaultT: T) {
     return { item, loading, error }
 }
 
-export default useFetchFromUrlT
+export default useInfiniteFetchFromUrlT
