+++ src/page/Themes.tsx
@@ -4,9 +4,10 @@ import {
 import ProductList from '@/component/theme/ProductList';
 import { BaseUrl } from '@/constant/api';
 import useFetchFromUrlT from '@/hook/useFetchFromUrlT';
-import { DefaultDiv, EmptyDiv4h, EmptyDiv8h } from '@/styles/CommomStyle/Common.styled';
+import { DefaultDiv, Gap } from '@/styles/CommomStyle/Common.styled';
 import { ThemeDescription, ThemeName, ThemeTitle, ThemeTop } from '@/styles/CommomStyle/themes.styled';
 import { defaultThemeInfo, type ThemeInfo } from '@/type/GiftAPI/product';
+import { getFromUrl } from '@/utils/getFromUrl';
 import { useParams } from 'react-router-dom';
 //import useFetchFromUrlT from '@/hook/useFetchFromUrlT';
 
@@ -15,7 +16,7 @@ import { useParams } from 'react-router-dom';
 const Themes = () => {
   const { themeId } = useParams<{ themeId: string }>();
   const themesUrl = `${BaseUrl}/api/themes/${themeId}/info`
-  const themeInfo = useFetchFromUrlT<ThemeInfo>(themesUrl,defaultThemeInfo);
+  const themeInfo = useFetchFromUrlT<ThemeInfo>(themesUrl, getFromUrl, defaultThemeInfo);
 
   const themeBackground = (themeInfo.item?.backgroundColor ?? 'white')
   const themeName = themeInfo.item?.name
@@ -27,9 +28,9 @@ const Themes = () => {
     <DefaultDiv>
       <ThemeTop background={themeBackground}>
         <ThemeName> {themeName} </ThemeName>
-        <EmptyDiv8h/>
+        <Gap height={8} />
         <ThemeTitle>{themeTitle}</ThemeTitle>
-        <EmptyDiv4h/>
+        <Gap height={4} />
         <ThemeDescription>{themeDescription}</ThemeDescription>
       </ThemeTop>
       <ProductList />
